update samples for redis cache
