+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache-passwordless/src/main/resources/application.yaml
@@ -1,0 +1,8 @@
+spring:
+  redis:
+    host: ${AZURE_CACHE_REDIS_HOST}
+    port: 6380
+    username: ${AZURE_CACHE_REDIS_USERNAME}
+    ssl: true
+    azure:
+      passwordless-enabled: true
