+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache-passwordless/terraform/setup_env.ps1
@@ -1,0 +1,3 @@
+$env:AZURE_CACHE_REDIS_HOST=$(terraform -chdir=terraform output -raw AZURE_CACHE_REDIS_HOST)
+
+echo AZURE_CACHE_REDIS_HOST=$env:AZURE_CACHE_REDIS_HOST
