+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache-passwordless/terraform/setup_env.sh
@@ -1,0 +1,3 @@
+export AZURE_CACHE_REDIS_HOST=$(terraform -chdir=./terraform output -raw AZURE_CACHE_REDIS_HOST)
+
+echo AZURE_CACHE_REDIS_HOST=AZURE_CACHE_REDIS_HOST
