+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache-passwordless/terraform/variables.tf
@@ -1,0 +1,17 @@
+variable "application_name" {
+  type        = string
+  description = "The name of your application."
+  default     = "cache-sample"
+}
+
+variable "location" {
+  type        = string
+  description = "The Azure region where all resources in this example should be created."
+  default     = "eastus"
+}
+
+variable "sample_tag_value" {
+  type        = string
+  description = "The value of spring-cloud-azure-sample tag."
+  default     = "true"
+}
