+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache/src/main/java/com/azure/spring/sample/cache/CacheApplication.java
