+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache/src/main/resources/application.yaml
@@ -1,0 +1,9 @@
+spring:
+  cloud:
+    azure:
+      profile:
+        subscription-id: ${AZURE_SUBSCRIPTION_ID}
+      redis:
+        name: ${AZURE_CACHE_REDIS_NAME}
+        resource:
+          resource-group: ${AZURE_RESOURCE_GROUP}
