+++ cache/spring2-sample/spring-cloud-azure-starter/spring-cloud-azure-sample-cache/terraform/setup_env.ps1
@@ -1,0 +1,7 @@
+$env:AZURE_SUBSCRIPTION_ID=$(terraform -chdir=terraform output -raw AZURE_SUBSCRIPTION_ID)
+$env:AZURE_RESOURCE_GROUP=$(terraform -chdir=terraform output -raw AZURE_RESOURCE_GROUP)
+$env:AZURE_CACHE_REDIS_NAME=$(terraform -chdir=terraform output -raw AZURE_CACHE_REDIS_NAME)
+
+echo AZURE_SUBSCRIPTION_ID=$env:AZURE_SUBSCRIPTION_ID
+echo AZURE_RESOURCE_GROUP=$env:AZURE_RESOURCE_GROUP
+echo AZURE_CACHE_REDIS_NAME=$env:AZURE_CACHE_REDIS_NAME
