+++ cache/spring3-sample/spring-cloud-azure-redis-sample-passwordless/src/main/java/com/azure/spring/sample/cache/CachePasswordlessApplication.java
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.spring.sample.cache;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+
+@SpringBootApplication
+@EnableCaching
+public class CachePasswordlessApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CachePasswordlessApplication.class, args);
+    }
+}
