+++ cache/spring3-sample/spring-cloud-azure-redis-sample-passwordless/src/main/java/com/azure/spring/sample/cache/WebController.java
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.spring.sample.cache;
+
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class WebController {
+
+    @GetMapping("/{name}")
+    @Cacheable("azureCache")
+    public String getValue(@PathVariable("name") String name) {
+        return "Hello " + name;
+    }
+}
