+++ cache/spring3-sample/spring-cloud-azure-redis-sample-passwordless/src/main/resources/application.yaml
@@ -1,0 +1,10 @@
+spring:
+  data:
+    redis:
+      host: ${AZURE_CACHE_REDIS_HOST}
+      port: 6380
+      username: ${AZURE_CACHE_REDIS_USERNAME}
+      ssl:
+        enabled: true
+      azure:
+        passwordless-enabled: true
