+++ cache/spring3-sample/spring-cloud-azure-redis-sample-passwordless/terraform/outputs.tf
@@ -1,0 +1,4 @@
+output "AZURE_CACHE_REDIS_HOST" {
+  value       = azurerm_redis_cache.redis.hostname
+  description = "The host name of the Redis instance."
+}
