+++ cache/spring3-sample/spring-cloud-azure-redis-sample-passwordless/terraform/variables.tf
