+++ cache/spring3-sample/spring-cloud-azure-redis-sample/src/main/resources/application.yaml
