+++ cache/spring3-sample/spring-cloud-azure-redis-sample/terraform/setup_env.sh
